Move backend tests to their own file
