+++ yjit/src/backend/mod.rs
@@ -1,3 +1,3 @@
 pub mod x86_64;
-
 pub mod ir;
+mod tests;
